Complete plan consolidation
