+++ openspec/specs/docs-cleanup/spec.md
@@ -2,9 +2,7 @@
 
 ## Purpose
 TBD - normalized from existing baseline requirements.
-
 ## Requirements
-
 ### Requirement: Repository markdown avoids stale local paths
 Tracked markdown documentation MUST NOT contain stale absolute local paths such as `/home/user/...`, except inside archived OpenSpec artifacts that intentionally preserve historical examples.
 
@@ -27,3 +25,19 @@ The package metadata MUST include at least ten relevant npm keywords, and the RE
 - **WHEN** the package manifest and README are inspected
 - **THEN** `package.json` has at least ten finance-relevant keywords
 - **AND** README contains a "Why OpenCandle?" heading
+
+### Requirement: Completed Generated Plans Archive Under OpenSpec
+
+Completed generated implementation-plan queues SHALL be preserved under archived OpenSpec changes instead of remaining as a root-level `plans/` folder.
+
+#### Scenario: Generated plan queue is complete
+
+- **WHEN** every item in a generated plan queue is implemented or intentionally rejected
+- **THEN** the queue is moved into a completed OpenSpec change as source evidence
+- **AND** the root of the repository no longer contains the generated `plans/` folder
+
+#### Scenario: OpenSpec remains the completed-work ledger
+
+- **WHEN** maintainers inspect completed generated implementation work
+- **THEN** they can find the proposal, completed tasks, and original plan records under `openspec/changes/archive/`
+- **AND** no second root-level planning index is required for those completed records
